+++ packages/vm/src/instance.rs
@@ -17,7 +17,7 @@ use crate::context::{
     move_into_context, move_out_of_context, setup_context, with_storage_from_context,
 };
 use crate::conversion::to_u32;
-use crate::errors::{VmResult, WasmerErr, WasmerRuntimeErr};
+use crate::errors::{make_instantiation_err, VmResult, WasmerRuntimeErr};
 use crate::features::required_features_from_wasmer_instance;
 use crate::imports::{
     do_canonicalize_address, do_humanize_address, do_query_chain, do_read, do_remove, do_write,
@@ -125,7 +125,9 @@ where
             },
         });
 
-        let wasmer_instance = module.instantiate(&import_obj).context(WasmerErr {})?;
+        let wasmer_instance = module.instantiate(&import_obj).map_err(|original| {
+            make_instantiation_err(format!("Error instantiating module: {:?}", original))
+        })?;
         Ok(Instance::from_wasmer(wasmer_instance, deps, gas_limit))
     }