+++ CHANGELOG.md
@@ -156,6 +156,7 @@
 - Remove `VmError::IoErr` in favour of `VmError::CacheErr`.
 - Simplify `VmError::CompileErr` and `VmError::ResolveErr` to just hold a string
   with the details instead of the source error.
+- Remove `VmError::WasmerErr` in favour of the new `VmError::InstantiationErr`.
 
 ## 0.7.2 (2020-03-23)